backend/requirements.txt
Django==3.2.3
django-cleanup==8.0.0
django-filter==23.2
django-sqlprint-middleware==0.1.4
djangorestframework==3.14.0
djoser==2.1.0
Pillow==10.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.3
PyYAML==6.0